# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "New Gate CRUD Application API",
    "description": "API with elasticsearch and mongodb",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/searchTask": {
      "post": {
        "tags": [
          "Task"
        ],
        "summary": "Search task name list",
        "description": "Search task name via ES",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to create",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New user is created",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        }
      }
    },
    "/getList": {
      "get": {
        "tags": [
          "Task"
        ],
        "summary": "Get all task in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        }
      }
    },
    "/modifyTask": {
      "post": {
        "tags": [
           "Task"
        ],
        "summary": "Modify task name",
        "description": "Find and modify task name",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to create",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Modify task is successfully",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        }
      }
    },
    "/deleteTask": {
      "post": {
        "tags": [
           "Task"
        ],
        "summary": "Delete task name",
        "description": "Delete task name",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "Delete task name",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Modify task is successfully",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Task": {
      "required": [
        "name"
      ],
      "properties": {
        "name": {
          "type": "string",
          "uniqueItems": true
        }
      }
    }
  }
}